Start moving code from game-core to game-headed
Starting with the entry-point to game-core. For now the moving of code
is pretty straightforward 1:1 moves.
